pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Zidane Wright <zidanewright@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
requests = "^2.31.0"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
jira = {extras = ["cli"], version = "^3.8.0"}
python-gitlab = "^4.4.0"
python-dotenv = "^1.0.1"
dacite = "^1.8.1"
distribute = "^0.7.3"
pytest-cov = "^5.0.0"
flake8 = "^7.0.0"
prometheus-fastapi-instrumentator = "^7.0.0"
autopep8 = "^2.1.0"
mutmut = "^2.4.5"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pre-commit = "^3.7.0"
tox = "^4.14.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"

[tool.pytest.ini_options]
markers = [
	"unit",
	"integration",
	"e2e"
]
